requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.11
python-dotenv==1.0.0
gunicorn==21.2.0
bcrypt==4.0.1
python-dateutil==2.8.2
Flask-Bcrypt==1.0.1
